move from poetry to uv and use src/ structure
