+++ pyproject.toml
@@ -1,22 +1,30 @@
-[tool.poetry]
+[project]
 name = "birthday-telegram-bot"
 version = "2.0.0-alpha"
 description = ""
-authors = ["orehzzz <olegzzz749@gmail.com>"]
-license = "MIT"
+authors = [{ name = "orehzzz", email = "olegzzz749@gmail.com" }]
+requires-python = "~=3.10.0"
 readme = "README.md"
+license = "MIT"
+dependencies = [
+    "peewee>=3.17.1,<4",
+    "python-telegram-bot[job-queue]~=20.8",
+    "pytz~=2024.1",
+    "psycopg2-binary>=2.9.9,<3",
+    "requests>=2.31.0,<3",
+    "cryptography>=39.0.1",
+]
 
-[tool.poetry.dependencies]
-python = "^3.10"
-peewee = "^3.17.1"
-python-telegram-bot = {extras = ["job-queue"], version = "^20.8"}
-pytz = "^2024.1"
-psycopg2-binary = "^2.9.9"
-requests = "^2.31.0"
+[dependency-groups]
+dev = ["black>=24.2.0,<25"]
 
-[tool.poetry.group.dev.dependencies]
-black = "^24.2.0"
+[project.scripts]
+birthday-bot = "src.birthday_bot:main"
 
 [build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+requires = ["uv_build>=0.9.3"]
+build-backend = "uv_build"
+
+[tool.uv.build-backend]
+module-name = "src"
+module-root = ""
