+++ src/birthdaybot_config_example.ini
