+++ src/core/logger.py
@@ -1,7 +1,6 @@
 import logging
-from logging.handlers import RotatingFileHandler
 import os
-
+from logging.handlers import RotatingFileHandler
 
 script_dir = os.path.dirname(os.path.abspath(__file__))
